+++ internal/cmd/docker/mediawiki/mediawiki.go
@@ -127,6 +127,7 @@ func NewMediaWikiCmd() *cobra.Command {
 	cmd.AddCommand(NewMediaWikiComposerCmd())
 	cmd.AddCommand(NewMediaWikiJobRunnerCmd())
 	cmd.AddCommand(NewMediaWikiExecCmd())
+	cmd.AddCommand(NewMediaWikiMWScriptCmd())
 	cmd.AddCommand(NewMediaWikiSitesCmd())
 	cmd.AddCommand(NewMediaWikiDoctorCmd())
 	return cmd